# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Scandere AI Store API",
        "description": "REST API for digital products e-commerce platform. Sell and manage digital documents (PDFs, Excel, Word) with secure downloads, payments via Stripe, and comprehensive admin panel.",
        "contact": {
            "name": "Scandere Support",
            "email": "team@scandere.info"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.scandereai.store",
            "description": "Production API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/products": {
            "get": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "List all products (admin)",
                "description": "Get paginated list of all products with category information. Requires admin authentication.",
                "operationId": "b34ae277f0fefc0a74d9869d25c1fa18",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated products list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (not admin)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Create new product (admin)",
                "description": "Create a new product with file upload. Requires admin authentication. Accepts multipart/form-data for file uploads.",
                "operationId": "ff11d12615a68edd8d75eb6fe4d5f885",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price",
                                    "file"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "New Product Template"
                                    },
                                    "category_id": {
                                        "description": "Category ID (nullable)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "short_description": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Brief product description"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Full product description with details"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number",
                                        "minimum": 0,
                                        "example": 9.99
                                    },
                                    "is_free": {
                                        "description": "Is this a free product",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "is_active": {
                                        "description": "Is product active (default: true)",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "is_featured": {
                                        "description": "Feature this product",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "file": {
                                        "description": "Product file (PDF, XLSX, DOCX, etc.) - max 50MB",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "preview_image": {
                                        "description": "Preview image (optional) - max 2MB",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (not admin)"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/{id}": {
            "get": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Get product details (admin)",
                "description": "Get detailed product information with category. Requires admin authentication.",
                "operationId": "062dc33bdb34eef6fb8eed7af62254c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (not admin)"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Update product (admin)",
                "description": "Update existing product. Optionally upload new file to replace existing. Requires admin authentication.",
                "operationId": "1234c087b21bd1df187ed757f67000ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "short_description": {
                                        "type": "string",
                                        "maxLength": 500
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "minimum": 0
                                    },
                                    "is_free": {
                                        "type": "boolean"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "is_featured": {
                                        "type": "boolean"
                                    },
                                    "file": {
                                        "description": "Optional: Replace product file (max 50MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (not admin)"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Delete product (admin)",
                "description": "Permanently delete product and associated file. Requires admin authentication.",
                "operationId": "07ca3329005fedce634d7afd90df29f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (not admin)"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format: Bearer {token}"
            }
        }
    },
...
}